Update README.md
Incorporate rbf121's method (per #389) of API requests to obtain memberslingbox info for slingboxes that were configured prior to the shutdown. Also, remove mention of the outdated method for the same process that no longer works after the shutdown.

I know that the rest of the document still needs a rework and various corrections, but this was a pretty quick and easy change to make.